requirements.txt
numpy
Pillow
matplotlib
tensorflow
csv2
